Cache resolve result on block
When cursor is on a symbol, an editor highlights all its usages. When
symbol is used a lot, the same name resolved again and
again (e.g. bits function in riscv.ion). The added one-element cache
remembers the last resolved result and reuses it if the same name is
resolved, similar to to what @pervognsen shown in one of the videos.